feat: 존재하지 않는 장바구니 아이템을 지우려 할 경우 예외처리 기능 구현

+++ src/main/java/cart/application/CartItemService.java
@@ -8,6 +8,7 @@ import cart.dto.request.CartItemQuantityUpdateRequest;
 import cart.dto.request.CartItemRequest;
 import cart.dto.response.CartItemResponse;
 import cart.exception.CartItemException.DuplicatedCartItem;
+import cart.exception.CartItemException.InvalidCartItem;
 import org.springframework.stereotype.Service;
 
 import java.util.Collections;
@@ -54,6 +55,9 @@ public class CartItemService {
 
     public void remove(Member member, Long id) {
         CartItem cartItem = cartItemDao.findById(id);
+        if (cartItem == null) {
+            throw new InvalidCartItem();
+        }
         cartItem.checkOwner(member);
 
         cartItemDao.deleteById(id);
